HW #05: Review fixes
